requirements.txt
# Core
anthropic
pydantic

# API
fastapi
uvicorn[standard]
websockets
pyzmq

# Database
sqlalchemy[asyncio]
aiosqlite
alembic

# Testing
pytest
pytest-asyncio
httpx